sumal edit 2025.07.24

+++ Model/SetupStyle.cs
@@ -1,6 +1,6 @@
 ﻿public class SetupStyle
 {
-    public int SetupStyleTypeID { get; set; }
+    public long SetupStyleTypeID { get; set; }
     public string SetupStyleCode { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public string? Remarks { get; set; }
