+++ Model/TravelAgent.cs
@@ -1,6 +1,6 @@
 ﻿public class TravelAgent
 {
-    public int TravelAgentID { get; set; }
+    public long TravelAgentID { get; set; }
     public string TravelAgentCode { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
